+++ entity/api_v1/views.py
@@ -1,4 +1,3 @@
-from airone.lib.profile import airone_profile
 from airone.lib.http import http_get
 from airone.lib.acl import ACLType
 
@@ -8,7 +7,6 @@ from django.http.response import JsonResponse
 from user.models import User
 
 
-@airone_profile
 @http_get
 def get_entities(request):
     user = User.objects.get(id=request.user.id)