+++ webhook/views.py
@@ -2,13 +2,11 @@ from airone.lib.acl import ACLType
 from airone.lib.http import get_object_with_check_permission
 from airone.lib.http import http_get
 from airone.lib.http import render
-from airone.lib.profile import airone_profile
 
 from entity.models import Entity
 from user.models import User
 
 
-@airone_profile
 @http_get
 def list_webhook(request, entity_id):
     user = User.objects.get(id=request.user.id)